Giving up on readline approach for now
I haven't been able to wrap readline for the python repl. I think it's
because Python has it's own logic for calling `dlopen` on things it
finds in `sys.path`, and if I use `LD_PRELOAD` to muck with that, it
works (somehow...), but I cannot then find the original readline
implementation by calling `dlsym(RTLD_NEXT, "...")`.

Ideas:
  1. Ship a complete readline implementation, rather than trying to wrap
     readline.
     This might actually not be too bad: the goal here is to support a
     very simple implementation: just feed a full line, no editing,
     completion, etc.
  2. Try wrapping libc's `read` function instead. The benefit of this is
     it would work in more scenarios (like someone reading from stdin
     themselves, rather than using readline). Might get weird as we're
     bypassing the kernel's TTY, though.

I'm going to explore 2) a bit. If that doesn't work out, I'll circle
back to 1).