Yet another approach, this time based on TTY ECHO
Unfortunately, the seccomp/ptrace approach didn't pan out on macOS.
AFAICT, you need god-mode privileges on macOS to do things like ptrace,
and I don't want to require that.

The approach here should be much more portable: when it looks like we're
getting prompted, we check to see if the TTY has `ECHO` enabled. If it's
not enabled, we assume it's ready for input, and we send a line of
input. If ECHO is disabled, then we don't send anything. This should
work for most (all?) REPLs, because they often use libraries like
`readline` and implement fancy features like history and tab completion,
which require disabling ECHO.